requirements.txt
certifi==2019.3.9
chardet==3.0.4
dj-static==0.0.6
Django==1.9.5
django-cors-headers==2.4.0
djangorestframework==3.5.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
Pillow==5.4.1
psycopg2==2.7.7
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
static3==0.7.0
urllib3==1.24.1
whitenoise==4.1.2
